+++ src/renocopro/policy/content/case_studies.py
@@ -1,16 +1,14 @@
 # -*- coding: utf-8 -*-
 
-from collective.taxonomy.interfaces import ITaxonomy
 from plone import schema
-from plone import api
 from plone.app.textfield import RichText
 from plone.dexterity.browser.view import DefaultView
 from plone.dexterity.content import Container
 from plone.supermodel.directives import fieldset
-from zope.component import getSiteManager
 from zope.interface import implements
 from renocopro.policy import _
 from renocopro.policy.content.interfaces import IRenocopro
+from renocopro.policy.utils import translate_selected_taxonomy_item
 
 
 class ICaseStudies(IRenocopro):
@@ -27,6 +25,7 @@ class ICaseStudies(IRenocopro):
             "location",
             "types_of_work",
             "type_of_building",
+            "construction_year",
             "renovation_year",
             "duration_of_the_work",
             "net_surface_area",
@@ -144,13 +143,5 @@ class CaseStudiesView(DefaultView):
             return True
         return False
 
-    def translate_selected_taxonomy_item(self, context, taxonomy_id, item_id):
-        sm = getSiteManager()
-        utility = sm.queryUtility(ITaxonomy, name=taxonomy_id)
-        if item_id:
-            if item_id:
-                return utility.translate(
-                    item_id,
-                    context=context,
-                    target_language=api.portal.get_current_language()[:2],
-                )
+    def get_taxonomy_item(self, context, taxonomy_id, item_id):
+        return translate_selected_taxonomy_item(context, taxonomy_id, item_id)